Extract TestGLContext to separate translation unit
For consistency with the Test.*Context classes for other backends, which
live in their own implementation file with their own header, extract
TestEGLContext to its own header and TU so that in cases where only a
TestEGLContext is required (e.g. EmbedderTestBackingStoreProducerGL), we
don't need to include all the various test GL surface classes as well.

Issue: https://github.com/flutter/flutter/issues/158998
