+++ testing/test_gl_surface.h
@@ -8,30 +8,11 @@
 #include <cstdint>
 
 #include "flutter/fml/macros.h"
-
+#include "flutter/testing/test_gl_context.h"
 #include "third_party/skia/include/core/SkSurface.h"
 #include "third_party/skia/include/gpu/ganesh/GrDirectContext.h"
 
 namespace flutter::testing {
-
-struct TestEGLContext {
-  explicit TestEGLContext();
-
-  ~TestEGLContext();
-
-  using EGLDisplay = void*;
-  using EGLContext = void*;
-  using EGLConfig = void*;
-
-  EGLDisplay display;
-  EGLContext onscreen_context;
-  EGLContext offscreen_context;
-
-  // EGLConfig is technically a property of the surfaces, no the context,
-  // but it's not that well separated in EGL (e.g. when
-  // EGL_KHR_no_config_context is not supported), so we just store it here.
-  EGLConfig config;
-};
 
 class TestGLOnscreenOnlySurface {
  public:
