+++ testing/test_gl_utils.h
@@ -1,0 +1,16 @@
+// Copyright 2013 The Flutter Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+#ifndef FLUTTER_TESTING_TEST_GL_UTILS_H_
+#define FLUTTER_TESTING_TEST_GL_UTILS_H_
+
+#include <string>
+
+namespace flutter::testing {
+
+std::string GetEGLError();
+
+}  // namespace flutter::testing
+
+#endif  // FLUTTER_TESTING_TEST_GL_UTILS_H_
